♻️ refactor: 인터페이스 설계 구조 리팩토링
